Fix the way match_pattern_base was applying mask
The edge_mask used to allow matching patterns near the edges of the
bitboard was applying only once but regardless the number of iteration
already done which can lead to having the the edge_mask in the resulting
bitboard.

This commit fixes that. Now the edge mask is applied under the same
conditions as before but only during the first iteration.