Rename get_win_move and make it use Node directly
This is done because Node now contains the last played move, which is
the move simulated by the said Node so, there is no need to bother
computing the move no more.

Also, the method itself was not named right as it doesn't give us the
winning move but only the next one to play.