Fix crash when AI plays first with DEPTH == 1
For the record, the issue was that we were comparing the `FScore` value
of the initial `Node` and the one returned by `get_next_move` which at
this stage in the game is very likely to be 0 in both `Node`s.

The fix is quite easy to do, what we have to do is simply to compare
the `BitBoard` `last_move` of each `Node`.